Move world generation from game entrypoint
