+++ src/game/main.ts
@@ -4,44 +4,18 @@ import { Engine } from '../engine/Engine'
 import { Experience } from '../engine/Experience'
 import { Resource } from '../engine/Resources'
 import { Benchmark } from './utilities/Benchmark'
-import { ChunkManager } from './world/ChunkManager'
-import { TerrainGenerator } from './world/TerrainGenerator'
-import { WorkerManager } from './world/workers/WorkerPool'
-import { ChunkMessageData } from './world/Chunk'
-import TerrainGenerationWorker from './world/workers/TerrainGenerationWorker.ts?worker'
+import { World } from './world/World'
 
 export default class Game implements Experience {
   resources: Resource[] = []
 
-  constructor(private engine: Engine) {}
+  constructor(private engine: Engine) { }
 
   @Benchmark
   init(): void {
-    const terrainGenerator = new TerrainGenerator(69420)
-    const chunkManager = new ChunkManager(terrainGenerator, new THREE.Vector3(8, 2, 8))
-
-    const chunks = chunkManager.createChunksAroundOrigin(new THREE.Vector3(0, 0, 0))
-
-    const workerCount = navigator.hardwareConcurrency
-
-    const workerManager = new WorkerManager<ChunkMessageData, ArrayBuffer>(
-      TerrainGenerationWorker,
-      workerCount
-    )
-
-    chunks.forEach((chunk) => {
-      this.engine.scene.add(chunk.mesh)
-
-      const task = chunk.generateTerrainGenerationWorkerTask()
-
-      workerManager.enqueueTask({
-        message: task.message,
-        callback: (args: MessageEvent<ArrayBuffer>) => {
-          task.callback(args)
-          chunk.updateMeshGeometry()
-        }
-      })
-    })
+    const world = new World(69420, new THREE.Vector3(8, 2, 8))
+    const chunks = world.generate()
+    chunks.forEach(chunk => this.engine.scene.add(chunk.mesh))
   }
 
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
